% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvalues.R
\name{mapvalues}
\alias{mapvalues}
\title{Map values}
\usage{
mapvalues(x, to, from, strict = F)
}
\arguments{
\item{x}{Vector to find map values to}

\item{to}{Vector of values to map to}

\item{from}{Vector of values to match}

\item{strict}{Logical. Strict type checking}
}
\value{
Vector with values mapped
}
\description{
Map values based on a from and to vector of equal length
}
\examples{
mapvalues(c(1,2,3), c(3,2,1), c(6,5,4))
}
