% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not in operator}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{Vector 1}

\item{y}{Vector 2}
}
\value{
Logical vector
}
\description{
Shows elements not in other vector
}
\examples{
c(1,2,3) \%!in\% c(1,3)
}
