% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{*}
\alias{*}
\title{Non-numeric multiplication}
\usage{
x * y
}
\arguments{
\item{x}{Non-numeric object}

\item{y}{Numeric}
}
\value{
A multiplicated string
}
\description{
Multiplicate non-numerics with the * operator
}
\examples{
"Hello World!" * 2
}
