% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vroom.R
\name{vroom}
\alias{vroom}
\title{Vroom wrapper with row_name functionality}
\usage{
vroom(file, delim = NULL, row_names = FALSE, ...)
}
\arguments{
\item{file}{File to open}

\item{delim}{Delimiter passed to vroom}

\item{row_names}{Logical. File has row.names (Default = FALSE)}
}
\value{
A tibble or data.frame
}
\description{
Use vroom::vroom() with added functionality for row_names
}
\examples{
data <- vroom("file.csv", row_names = T)
}
